requirement_latest.txt
absl-py==2.1.0
aiohttp==3.8.1
aiomysql==0.2.0
aiosignal==1.2.0
alembic==1.13.2
amqp==5.1.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
apispec==0.39.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.2.0
Babel==2.12.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
billiard==4.1.0
black==23.3.0
bleach==6.1.0
blinker==1.7.0
Brlapi==0.8.3
bs4==0.0.2
cachelib==0.9.0
cachetools==5.3.3
celery==5.3.1
certifi==2023.5.7
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
colorlover==0.3.0
comm==0.2.2
command-not-found==0.3
confluent-kafka==2.5.3
contourpy==1.3.0
cron-descriptor==1.4.3
cryptography==3.4.8
cufflinks==0.17.3
cupshelpers==1.0
cycler==0.12.1
Cython==0.29.35
dbus-python==1.2.18
debugpy==1.8.5
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-filter==24.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
elastic-transport==8.15.0
elasticsearch==7.8.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
execnet==1.9.0
executing==2.1.0
fake-useragent==1.5.1
fastapi==0.115.0
fastapi-cli==0.0.5
fasteners==0.14.1
fastjsonschema==2.20.0
filelock==3.6.0
flake8==6.0.0
Flask==3.0.0
flask-apispec==0.7.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.2.0
future==0.18.2
gast==0.6.0
-e git+ssh://git@github.com/CACF/GDDS_BACKEND.git@ec46850f8aa794d9cada887fb79f26f8b7fd12b3#egg=gdds
google-api-python-client==1.7.11
google-auth==2.29.0
google-auth-httplib2==0.1.0
google-pasta==0.2.0
graphene==3.3
graphene-sqlalchemy==2.1.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.1
grpcio==1.66.2
gyp==0.1
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.5.0
keyring==23.5.0
kiwisolver==1.4.7
kombu==5.3.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
limits==3.13.0
lockfile==0.12.2
louis==3.20.0
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
monotonic==1.6
more-itertools==9.1.0
multidict==5.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
mysqlclient==2.2.4
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
nlp==0.4.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
openai==1.35.3
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.12.1
ordered-set==4.1.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
paginate==0.5.6
pandas==2.0.2
pandas-datareader==0.10.0
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
pathspec==0.11.1
peewee==3.17.6
pexpect==4.8.0
pg8000==1.30.4
pika==0.13.0
Pillow==9.0.1
platformdirs==3.6.0
plotly==5.24.0
pluggy==1.2.0
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==4.25.5
proton-core==0.2.0
proton-keyring-linux==0.0.2
proton-keyring-linux-secretservice==0.0.2
proton-vpn-api-core==0.30.0
proton-vpn-gtk-app==4.4.3
proton-vpn-killswitch-network-manager==0.5.0
proton-vpn-killswitch-network-manager-wireguard==0.1.0
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.5.2
proton-vpn-network-manager-openvpn==0.1.0
proton-vpn-network-manager-wireguard==0.1.0
psutil==6.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
PyAMQP==0.1.0.7
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.22
pycups==2.0.1
pydantic==2.7.4
pydantic_core==2.18.4
pydash==7.0.3
PyDrive==1.3.1
pyflakes==3.0.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
pymongo==4.7.2
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.9.0
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.4.0
pytest-forked==1.2.0
python-apt==2.4.0+ubuntu3
python-crontab==3.0.0
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-json-logger==2.0.7
python-magic==0.4.15
python-multipart==0.0.9
pytz==2023.3
pyxdg==0.27
PyYAML==6.0
pyzmq==26.2.0
redis==4.6.0
referencing==0.35.1
reportlab==3.6.8
requests==2.31.0
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
Rx==1.6.3
scikit-learn==1.5.2
scipy==1.14.1
scramp==1.4.4
seaborn==0.13.2
SecretStorage==3.3.1
selenium==4.18.1
Send2Trash==1.8.3
sentry-sdk==1.4.3
shellingham==1.5.4
simplejwt==2.0.1
singledispatch==4.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
speaklater==1.3
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.38.5
systemd-python==234
tabulate==0.8.9
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.3.0
terminado==0.18.1
testing.common.database==2.0.3
texttable==1.6.4
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.5
types-python-dateutil==2.9.0.20240906
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
unicorn==2.0.1.post1
uri-template==1.3.0
uritemplate==3.0.1
urllib3==2.0.2
usb-creator==0.3.7
uvicorn==0.30.6
uvloop==0.20.0
vine==5.0.0
virtualenv==20.13.0+ds
virtualenvwrapper-win==1.2.7
wadllib==1.3.6
watchfiles==0.24.0
wcwidth==0.2.6
webargs==8.3.0
webcolors==24.8.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.1
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
xxhash==3.5.0
yarl==1.7.2
yfinance==0.2.43
zipp==3.15.0
